use is_last_message_from_operator
